refactor: big change in backend directories structure - switch to modular
