+++ backend/app/modules/admin/seed/router.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter(prefix="/seed", tags=["Admin seeding"])
+
+
+@router.post("/seed")
+def seed_database():
+    pass
