+++ backend/app/modules/events/schemas.py
@@ -1,17 +1,11 @@
 from datetime import datetime
 from enum import Enum
-from typing import Any, List, Optional
+from typing import List, Optional
 
 from pydantic import BaseModel, ConfigDict
 
-
-class FieldType(str, Enum):
-    string = "string"
-    number = "number"
-    integer = "integer"
-    boolean = "boolean"
-    array = "array"
-    object = "object"
+from app.modules.fields.schemas import FieldOut
+from app.modules.tags.schemas import TagOut
 
 
 class LinkType(str, Enum):
@@ -29,45 +23,6 @@ class EventLink(BaseModel):
     type: LinkType
     url: str
     label: Optional[str] = None
-
-
-class TagBase(BaseModel):
-    id: str
-    description: str | None = None
-
-
-class TagCreate(TagBase):
-    pass
-
-
-class TagOut(TagBase):
-    created_at: datetime
-    updated_at: datetime
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class FieldBase(BaseModel):
-    name: str
-    description: Optional[str] = None
-    field_type: FieldType
-    example: Optional[Any] = None
-
-
-class FieldCreate(FieldBase):
-    pass
-
-
-class FieldOut(FieldBase):
-    id: int
-    created_at: datetime
-    updated_at: datetime
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class FieldOutWithEventCount(FieldOut):
-    event_count: int
 
 
 class EventBase(BaseModel):
